* rename loadtool -> bbencheth * `bbencheth load` is equivelent to the old loadtool: it generates load   and optionally collects results * `bbench collect` just collects the results * in anticipation of adding further sub commands in bbencheth, sort out   the cobra/vipra machinery.
The motivation for this split is support for generating in cluster loads
using kubernetes jobs. emptyDir is not accessible after the job
completes and I would rather not mess around with job specific
persistent volumes. As the data for the chainhamer compatible graphs is
always read from the blocks we can generate the load in cluster and read
the blocks (to get the results) wherever (and whenever) we like.