nodeport working.
not completely settled on whether to use --nat=extip:PUBLICIP for the
bootnodes. It's presence seems to make peering happen faster. But it
also prevents the boot nodes from peering with each other due to the
Lan/Wan restriction on local vs remote ports when nat of any kind is
enabled